Update dependencies and Kotlin JVM target
This commit updates the following dependencies to their latest versions:
- Android Gradle Plugin (AGP) from 8.10.1 to 8.11.0
- Kotlin from 2.1.21 to 2.2.0
- Coil Video from 3.1.0 to 3.2.0
- AppToolkit from 1.0.34 to 1.0.35

Additionally, it refactors the Kotlin JVM target configuration in `app/build.gradle.kts` to use the `kotlin.compilerOptions` block for setting `jvmTarget` to `JvmTarget.JVM_21`.

Finally, a minor adjustment was made in `MainScreen.kt` to pass an empty `PaddingValues()` to `AppNavigationHost` within the tablet layout to ensure consistent padding.